% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getAgricultureDepartments}
\alias{getAgricultureDepartments}
\title{Agriculture Departments Code}
\usage{
getAgricultureDepartments(api_key)
}
\arguments{
\item{api_key}{Your API application key}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item sectionCode
 \item sectionName
}
}
\description{
Agriculture departments code
}
\examples{
\dontrun{
 getAgricultureDepartments(api_key)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/regions/agricultureDepartments.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
