% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getBroadIndustries}
\alias{getBroadIndustries}
\title{Broad Industry Category}
\usage{
getBroadIndustries(api_key)
}
\arguments{
\item{api_key}{Your API application key}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item sicCode
 \item sicName
}
}
\description{
List of broad industry category codes
}
\examples{
\dontrun{
 getBroadIndustries(api_key)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/industries/broad.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
