% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getBroadJobs}
\alias{getBroadJobs}
\title{Broad Job Category}
\usage{
getBroadJobs(api_key)
}
\arguments{
\item{api_key}{Your API application key}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item iscoCode
 \item iscoName
}
}
\description{
Broad job category codes
}
\examples{
\dontrun{
 getBroadJobs(api_key)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/jobs/broad.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
