% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getCities}
\alias{getCities}
\title{List of Cities}
\usage{
getCities(api_key, prefCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{prefCode}{Code of a prefecture}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item prefCode
 \item cityCode
 \item cityName
 \item bigCityFlag
}
}
\description{
Cities' codes
}
\examples{
\dontrun{
 getCities(api_key, prefCode = 1)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/cities.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
