% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCityPrefCode.R
\name{getCityPrefCode}
\alias{getCityPrefCode}
\title{City Code}
\usage{
getCityPrefCode(api_key, pref = "東京都")
}
\arguments{
\item{api_key}{Your API application key}

\item{pref}{Name of prefecture to fetch information for}
}
\value{
This function returns \code{date.fram} including columns:
\itemize{
 \item prefCode
 \item cityCode
 \item cityName
 \item bigCityFlag
}
}
\description{
Get list of city codes of a given prefecture
}
\examples{
\dontrun{
  getCityPrefCode(api_key='Your API key', pref='東京都')
}

}
\author{
Koki Ando
}
