% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmploymentStats.R
\name{getEmploymentStat}
\alias{getEmploymentStat}
\title{Employment Data}
\usage{
getEmploymentStat(api_key, prefCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{prefCode}{Code of a prefecture to fetch information for}
}
\description{
Data of employment data of a given prefecture in Japan
}
\examples{
\dontrun{
 getEmploymentStat(api_key, prefCode = 26)
}

}
\author{
Koki Ando
}
