% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourism.R
\name{getForeignTourists}
\alias{getForeignTourists}
\title{Foreigners visiting specified place}
\usage{
getForeignTourists(api_key, year, prefCode, purpose)
}
\arguments{
\item{api_key}{Your API application key}

\item{year}{Number of year}

\item{prefCode}{Code of a prefectrue}
}
\value{
This function returns \code{data.fram} including columns:
\itemize{
 \item country
 \item year
 \item quater
 \item value
}
}
\description{
The number of foreigners visiting a specified place per year
}
\examples{
\dontrun{
 getForeignTourists(api_key, year = "2012", prefCode = "11")
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/tourism/foreigners/forFrom.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
