% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourism.R
\name{getForeiners}
\alias{getForeiners}
\title{Foreigners visiting specified place}
\usage{
getForeiners(api_key, year, prefCode, purpose)
}
\arguments{
\item{api_key}{Your API application key}

\item{purpose}{}
}
\description{
The number of foreigners visiting a specified place per year
}
\examples{
\dontrun{
 get(api_key, year = "2012", prefCode = "11", purpose= "1")
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/tourism/foreigners/forFrom.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
