% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourism.R
\name{getGuestsNum}
\alias{getGuestsNum}
\title{Number of guests}
\usage{
getGuestsNum(api_key, unit, prefCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{unit}{Unit: 0 returns annual data, 1 returns monthly data}

\item{prefCode}{Code of a given prefecture}
}
\description{
Number of guests stayed at different types of accomodations in a given prefecture from 2011 to 2017
}
\examples{
\dontrun{
 getGuestsNum(api_key, unit = 1, prefCode = 1)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/tourism/hotelAnalysis/groupStack.html}
}
