% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourism.R
\name{getHotelOperationRate}
\alias{getHotelOperationRate}
\title{Operation Rate}
\usage{
getHotelOperationRate(api_key, unit, prefCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{unit}{Unit 0 returns annual data, 1 returns monthly data}

\item{prefCode}{Code of prefecture}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item type
 \item year
 \item month
 \item value
 \item prefCode
}
}
\description{
Operation rata data for multiple types of accomodation of a given prefecture from 2011 to 2017
}
\examples{
\dontrun{
 getHotelOperationRate(api_key, unit = 0, prefCode = 1)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/tourism/hotelAnalysis/guestRoomLine.html}
}
\author{
Koki Ando <koki.25.andog@gmail.com>
}
