% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndustryCategories.R
\name{getIndustryJobCode}
\alias{getIndustryJobCode}
\title{Indusry or Job Code}
\usage{
getIndustryJobCode(api_key, category = c("産業", "職業"),
  size = c("大", "中", "小"))
}
\arguments{
\item{api_key}{Your API application key}

\item{category}{Category of code. valid values are 職業 or 産業}

\item{size}{Size of code. Valid values are 大, 中 or 小}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item iscoCode
 \item iscoName
}
}
\description{
Get data of industry or job codes
}
\examples{
\dontrun{
  getIndustryJobCode(api_key="*******", 
                     category = "職業", size = "大")
}

}
\author{
Koki Ando
}
