% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{getLocalTexes}
\alias{getLocalTexes}
\title{Local taxes per year}
\usage{
getLocalTexes()
}
\arguments{
\item{api_key}{Your API application key}

\item{prefCode}{Code of prefecture}

\item{cityCode}{Code of city}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item year
 \item value
 \item prefCode
 \item pref
 \item cityCode
 \item city
}
}
\description{
Local taxes (thousands) per capita per year of specific municipality
}
\examples{
\dontrun{
 getLocalTexes(api_key, prefCode=11, cityCode=11362)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/municipality/taxes/perYear.html}
}
\author{
Koki Ando <koki.25.ando@gmail.co>
}
