% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getMiddleIndustries}
\alias{getMiddleIndustries}
\title{Middle Industry Category}
\usage{
getMiddleIndustries(api_key, sicCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{sicCode}{Industry broad code}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item simcCode
 \item sicCode
 \item simcName
}
}
\description{
Middle industry category codes
}
\examples{
\dontrun{
 getMiddleIndustries(api_key, sicCode = "E")
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/industries/middle.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
