% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getMiddleJobs}
\alias{getMiddleJobs}
\title{Middle Job Category}
\usage{
getMiddleJobs(api_key, iscoCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{iscoCode}{Code of broad job}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item iscoCode
 \item ismcoCode
 \item iscoName
}
}
\description{
Middle job category codes
}
\examples{
\dontrun{
 getMiddleJobs(api_key, iscoCode = "B")
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/jobs/middle.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
