% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getNarroaIndustries}
\alias{getNarroaIndustries}
\title{Narrow Industry Categories}
\usage{
getNarroaIndustries(api_key, simcCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{simcCode}{Industry middle code}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item siscCode
 \item simcCode
 \item sidcName
}
}
\description{
Narrow industry category codes
}
\examples{
\dontrun{
 getNarroaIndustries(api_key, simcCode = 20)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/industries/narrow.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
