% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{getPopulationChange}
\alias{getPopulationChange}
\title{Population Change}
\usage{
getPopulationChange(api_key, prefCode, cityCode = "-")
}
\arguments{
\item{api_key}{Your API application key}

\item{prefCode}{Code of a prefecture}

\item{cityCode}{Code of a city. "-" return all the cities}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item label
 \item value
 \item year
 \item prefCode
}
}
\description{
Population change data every 5 years from 1985 to 2040
}
\examples{
\dontrun{
 getPopulationChange(api_key, prefCode = 1, cityCode = "-")
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/population/sum/perYear.html}
}
\author{
Koki Ando <koki.25.ando@gmail.gmail>
}
