% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{getPopulationComposition}
\alias{getPopulationComposition}
\title{Population Composition}
\usage{
getPopulationComposition(api_key, prefCode, cityCode = "-")
}
\arguments{
\item{api_key}{Your API application key}

\item{prefCode}{Code of a prefecture}

\item{cityCode}{Code of a city. "-" returns all the cities}
}
\description{
Population composition data of a given prefecture in 2015
}
\examples{
\dontrun{
 getPopulationComposition(api_key, prefCode = 1, cityCode = "-")
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/population/composition/perYear.html}
}
\author{
Koki Ando <koki.25.ando@gmail.gmail>
}
