% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{getPopulationPyramid}
\alias{getPopulationPyramid}
\title{Population Pyramid}
\usage{
getPopulationPyramid(api_key, cityCode = "-", prefCode, yearLeft,
  yearRight)
}
\arguments{
\item{api_key}{Your API application key}

\item{cityCode}{Code of a city. "-" returns all the cities}

\item{prefCode}{Code of a prefecture}

\item{yearLeft}{Year number 1}

\item{yearRight}{Year number 2}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item class
 \item man
 \item manPercent
 \item woman
 \item womanPercent
 \item year
 \item type
 \item prefCode
}
}
\description{
Population pyramid data of a given prefecture and municipality in a 2 different years
}
\examples{
\dontrun{
 getPopulationPyramid(api_key, cityCode, prefCode = 1, yearLeft = 2020, yearRight = 2030)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/population/composition/pyramid.html}
}
\author{
Koki Ando <koki.25.ando@gmail.gmail>
}
