% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopulationStats.R
\name{getPopulationStats}
\alias{getPopulationStats}
\title{Population data for a given place}
\usage{
getPopulationStats(api_key, cityCode, prefCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{cityCode}{City codes. You can select all cities by using "-". You can check these
codes by using \code{getCityPrefCode} function.}

\item{prefCode}{Prefecture codes.}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item 年
 \item 総人口
 \item 年少人口
 \item 年少人口比率
 \item 生産年齢人口
 \item 生産年齢人口比率
 \item 老年人口
 \item 老年人口比率
 \item cityCode
 \item prefCode
}
}
\description{
This function provides population data in a given city and prefecture
}
\examples{
\dontrun{
  getPopulationStats(api_key, cityCode=26100, prefCode=26)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/population/composition/perYear.html}
}
\author{
Koki Ando
}
