% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{getPrefectures}
\alias{getPrefectures}
\title{List of Prefectures}
\usage{
getPrefectures(api_key)
}
\arguments{
\item{api_key}{Your API application key}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item prefCode
 \item prefName
}
}
\description{
Prefectures' codes
}
\examples{
\dontrun{
 getPrefectures(api_key)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/prefectures.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
