% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{getPropertyTax}
\alias{getPropertyTax}
\title{Property Tax}
\usage{
getPropertyTax(api_key, cityCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{cityCode}{Code of city}
}
\description{
Property tax per capita per year of a given municipality from 2008 to 2016
}
\examples{
\dontrun{
 getPropertyTax(api_key, cityCode = 11362)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/municipality/propertyTax/perYear.html}
}
