% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipality.R
\name{getSettlementAmount}
\alias{getSettlementAmount}
\title{Fiancial information of specific municipality}
\usage{
getSettlementAmount(api_key, year, prefCode, cityCode, matter = 1)
}
\arguments{
\item{api_key}{Your API application key}

\item{year}{Specific number of fiscal year}

\item{prefCode}{Code of prefecture}

\item{cityCode}{Code of municipality}

\item{matter}{Specified are code}
}
\value{
This function returns time series \code{data.frame} including columns: 
\itemize{
 \item label
 \item value
 \item year
 \item prefCode
 \item pref
 \item cityCode
 \item city
}
}
\description{
Settlement amount ratio of each category for a municipality
}
\examples{
\dontrun{
 getSettlementAmount(api_key, year = 2015, prefCode=26, cityCode=26100)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/municipality/finance/forSettlementAmount.html}
}
\author{
Koki Ando <koki.25.ando@gmail.com>
}
