% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTourismStats.R
\name{getTourismStats}
\alias{getTourismStats}
\title{The number of foreigners who visited a given prefecture}
\usage{
getTourismStats(api_key, purpose = 1, year = 2012, prefCode)
}
\arguments{
\item{api_key}{Your API application key}

\item{purpose}{Purpose of visiting. Defaults 1. 1 means "ALL", 2 means "Sightseeing or Leisure".}

\item{year}{Year. Years from 2011 to 2016 are only valid. Defaults 2012.}

\item{prefCode}{Prefecture code}
}
\value{
This function returns \code{data.frame} including columns:
\itemize{
 \item year
 \item quarter
 \item value
 \item Country
}
}
\description{
This function provides the number of foreigners who visited a given prefecture. 
Also provides where they are from and when they visited.
}
\examples{
\dontrun{
  getTourismStats(api_key, year = 2013, prefCode=26)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/tourism/foreigners/forFrom.html}
}
\author{
Koki Ando
}
