% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{getWages}
\alias{getWages}
\title{Wages per capita per year}
\usage{
getWages(api_key, prefCode, wagesAge = 1)
}
\arguments{
\item{api_key}{Your API application key}

\item{prefCode}{Code of prefecture}
}
\description{
Wages per capita per year of a given prefecture
}
\examples{
\dontrun{
 getWages(api_key, prefCode = 11, wagesAge = 1)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/municipality/wages/perYear.html}
}
