% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResasAPI.R
\docType{package}
\name{ResasAPI}
\alias{ResasAPI}
\alias{ResasAPI-package}
\title{ResasAPI: Access to RESAS API
Provides an interface to the RESAS API and allows R users to download piblicly available data 
on RESAS API.}
\description{
What the package does (one paragraph).
}
\seealso{
\url{https://opendata.resas-portal.go.jp/}
}
\author{
Koki Ando \email{koki.25.ando@gmail.com}
}
