% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourism.R
\name{getFacilityNum}
\alias{getFacilityNum}
\title{Facilities of a prefecture}
\usage{
getFacilityNum(api_key, display, unit, prefCode)
}
\description{
The number of facilities in a given prefecture per year
}
\examples{
\dontrun{
 getFacilityNum(api_key, display = 1, unit = 0, prefCode = 1)
}

}
\seealso{
\url{https://opendata.resas-portal.go.jp/docs/api/v1/tourism/hotelAnalysis/facilityStack.html}
}
